moves ConditionScores from conditions to plan
